Created CsrfToken class
